refactor(controllers): refactor controllers of API routes to make code more modular
