--- server/controllers/diary controller/getSingleDiaryEntryController.js
@@ -1,45 +1,0 @@
-import dummyDatabase from '../../data structures/dummyDatabase';
-import validIdValidator from '../../validators/validIdValidator';
-/**
- * @class diarycontroller
- */
-class diarycontroller {
-  /**
-     * @description get single diary entry
-     * @param {*} req http request
-     * @param {*} res http response
-     * @returns {JSON} returns a JSON object
-     */
-  static getSingleEntry(req, res) {
-    // Get id from params
-    const { id } = req.params;
-
-    // Validate the id and check if it's an integer - change to check for UUID later
-    // remember to write test reflecting use of UUID later
-    const { errors, isValid } = validIdValidator(req.params);
-
-    // id from params is string and is converted to an integer for comparison purposes
-    // filter each dummy database element based on id
-    const selectedEntry = dummyDatabase.filter(entry => parseInt(id, 0) === entry.id);
-
-    // if no validation errors, check that data exists in database
-    // if data exists, give ok status and data
-    if (isValid && dummyDatabase.length > 0 && selectedEntry.length === 1) {
-      res.status(200).json({
-        data: selectedEntry[0]
-      });
-    // if data doesn't exist, give 404 status and error message of not found
-    } else if (isValid && dummyDatabase.length > 0 && selectedEntry.length === 0) {
-      res.status(404).json({
-        message: 'Diary entry not found'
-      });
-      // if validation error, give error message
-    } else {
-      res.status(400).json({
-        errors: errors.id
-      });
-    }
-  }
-}
-
-export default diarycontroller;
